[수정] tradeAmount Double로 타입변경

+++ src/main/java/com/AiFunding/ToBi/dto/trade/signal/TradeSignalDetailDto.java
@@ -11,9 +11,9 @@ public class TradeSignalDetailDto {
 
     private Integer tradeSignal;
 
-    private Integer tradeAmount;
+    private Double tradeAmount;
 
-    TradeSignalDetailDto(final String stockName, final Integer tradeSignal, final Integer tradeAmount){
+    TradeSignalDetailDto(final String stockName, final Integer tradeSignal, final Double tradeAmount){
         this.stockName = stockName;
         this.tradeAmount = tradeAmount;
         this.tradeSignal = tradeSignal;
